+++ pkg/engine/attestation_test.go
@@ -6,6 +6,7 @@ import (
 
 	v1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	"github.com/kyverno/kyverno/pkg/engine/context"
+	"github.com/kyverno/kyverno/pkg/engine/internal"
 	"github.com/kyverno/kyverno/pkg/logging"
 	"github.com/kyverno/kyverno/pkg/utils/api"
 	"github.com/kyverno/kyverno/pkg/utils/image"
@@ -257,7 +258,7 @@ func Test_Conditions(t *testing.T) {
 	err := json.Unmarshal([]byte(scanPredicate), &dataMap)
 	assert.NilError(t, err)
 
-	pass, err := evaluateConditions(conditions, ctx, dataMap, logging.GlobalLogger())
+	pass, err := internal.EvaluateConditions(conditions, ctx, dataMap, logging.GlobalLogger())
 	assert.NilError(t, err)
 	assert.Equal(t, pass, true)
 }
